Babeltrace python binding
The Babeltrace Python module can be used to directly control
the Babeltrace API inside Python, using 'import babeltrace'.

Therefore, it becomes possible to create a Context, add a
trace to it, iterate on it, read events and so on from
within Python.

SWIG >= 2.0 is used to create the wrapper and its
'warning md variable unused' bug is patched in Makefile.am

In the interface file, struct and enum are directly copied
from the include files. All changes to struct bt_iter_pos
and to enums in ctf/events.h and clock-types.h  must also
be made in the interface file.

Signed-off-by: Danny Serres <danny.serres@efficios.com>
Signed-off-by: Yannick Brosseau <yannick.brosseau@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>