Memstream functions for MinGW32
In Windows, you cannot unlink a file while it is still in use,
so we won't even try.
Add a Windows-specific implementation of mkstemp, which provides a
temporary file in the correct Windows directory. The temporary
file is opened with such flags that it will be deleted once it's closed.