Add a MinGW substitute for openat() and related functions
Check if the build system has openat() support.
If it does not, we have to open files based on the directory name
instead of the directory handle. Substitute openat() with
compat_openat() that also takes the directory path name as a
parameter.
If we don't have openat(), replace the function dirfd() with a dummy.
Also, if we don't have openat(), open() and close() will not
work for directories. Replace those with compat_opendirfd and
compat_closedirfd() with dummy equivalents, as the function
compat_openat() takes care of opening the file.