Fix: Compound type write functions must check the offset before aligning
The ctf_access_ok() check was done after aligning. The write functions would
create a new packet if the ctf_pos_align could not align inside the current
packet.

The functions now report an error if the alignment can't be ensured in the
current packet. The top-level CTF Writer serialization functions will resize
the packets appropriately.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>