ir: allow the creation of an empty clock (nameless)
bt_ctf_clock_create_empty() is used to create an empty clock object.
This clock has no name, and it cannot be added to any other object
until its name is set using bt_ctf_clock_set_name().

This helps the creation of a clock object when its name is not known
in advance, just like the other properties.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>