ir: make bt_ctf_field_type_copy() public
This function can be useful for some source plugins which support
aliases in their metadata description: the base type is created
once, then one copy is created for each instance (possibly with
a different tag if it's a variant alias, for example).

The copy and all its children have their valid flag reset to 0 (as
well as their frozen flag). This means that they are not considered
valid after the copy operation. The length and tag field paths of
resp. sequence and variant field types will be resolved later when
this copy is used as one of the types of an event class, a stream
class, or a trace, and when it's about to be frozen in those
contexts.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>