Add Doxygen support to build system to build the API doc
This patch adds what is needed to generate the HTML documentation
of Babeltrace's C API using Doxygen to look for documentation source
in selected header files.

The new configure option --enable-api-doc enables the API doc
build using Doxygen (Doxygen must be found in this case). Whether or
not this API doc is built is indicated in the summary at the end of
the configure script execution.

Doxygen's input is doc/api/Doxyfile.in. This file has a few
substitutions like the project's version and the paths to input
files.

The API doc target, if enabled with the configure option, is part of
the `all` target.

With the `install` target, the files are installed in
$(DESTDIR)$(docdir)/api, e.g. /usr/share/doc/babeltrace/api.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>