babeltrace-cfg: add begin/end NS options
--base-begin-ns and --base-end-ns are similar to the --base-params
option, but they control the current base beginning and end timestamp.

--begin-ns and --end-ns are similar to the --params option, but they
assign a beginning/end timestamp to the latest source component
instance.

--reset-base-begin-ns and --reset-base-end-ns are similar to the
--reset-base-params option, but they reset the current base beginning
and end timestamps.

If both beginning and end timestamps are set for a given source
component instance, they are validated so that the end is greater
than or equal to the beginning.

This patch also reorders the options in the help message by long
option name.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>