Make bt_graph_add_*_listener() function return a listener ID
This is the expected API to identify all the user listeners of the
library. You can keep the returned ID, a simple int, to remove the
listener from the object later if needed. With the current API, you
cannot remove graph listeners, but we could add the functions to do so
in the future.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>