Add bt_common_normalize_path()
This new function does a part of realpath()'s job: it resolves `..` and
`.` in the path, resolves a relative path with a custom working
directory or the current working directory, and removes consecutive
separators (`/` or `\` depending on the platform) and the last
separator. The function always returns an absolute path. The function
does NOT resolve symbolic links.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>