ctf.fs source: use last path parameter's element as base for trace names
Without this patch:

    Trace directory:  /home/user/my-traces/trace/yes/kernel
    `path` parameter: /home/user/my-traces/trace
    Trace name:                                      kernel

With this patch:

    Trace directory:  /home/user/my-traces/trace/yes/kernel
    `path` parameter: /home/user/my-traces/trace
    Trace name:                            trace/yes/kernel

This takes into account absolute vs. relative paths, resolves `..` and
`.`, and more. Most of the job is done by bt_common_normalize_path().
The process used to craft the trace names using the base directory as
passed by the user is really easy once all the paths are known to be
normalized.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>