ctf: notif-iter.c: lazily initialize clock value states
Because clock classes can be added to a trace at any time, it is
possible that an integer field type is mapped to a clock class which did
not exist when init_clock_states() is initially called. To fix this,
create the clock state and map it to a clock class when a new clock
class is found.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>